+++ src/main/java/org/aksw/word2vecrestful/word2vec/W2VNrmlMemModelKMeans.java
@@ -20,11 +20,17 @@ import org.apache.log4j.Logger;
 import com.opencsv.CSVReader;
 import com.opencsv.CSVWriter;
 
-import nikit.test.TimeLogger;
-
 /**
  * Class to encapsulate word2vec in-memory model and expose methods to perform
- * search on the model
+ * search on the model.
+ * 
+ * This class selects {@link W2VNrmlMemModelKMeans#compareVecCount} vectors
+ * (centroids of the KMeans result on the model vectors) and then calculates the
+ * cosine similarity of all words in model to those vectors.
+ * 
+ * It uses the knowledge about pre-processed similarities with
+ * {@link W2VNrmlMemModelKMeans#comparisonVecs} to narrow down the search of
+ * closest word for the user specified vector.
  * 
  * @author Nikit
  *
@@ -42,8 +48,6 @@ public class W2VNrmlMemModelKMeans implements GenWord2VecModel {
 	private int kMeansMaxItr = 5;
 	private BitSet[][] csBucketContainer;
 	private String vecFilePath = "data/kmeans/comparison-vecs.csv";
-	// TODO : Remove this
-	private TimeLogger tl = new TimeLogger();
 
 	public W2VNrmlMemModelKMeans(final Map<String, float[]> word2vec, final int vectorSize) throws IOException {
 		this.word2vec = word2vec;
@@ -175,14 +179,13 @@ public class W2VNrmlMemModelKMeans implements GenWord2VecModel {
 			// calculate cosine similarity of all distances
 			float[] curCompVec;
 			BitSet finBitSet = null;
-			tl.logTime(1);
 			for (int i = 0; i < compareVecCount; i++) {
 				curCompVec = comparisonVecs[i];
 				double cosSimVal = Word2VecMath.cosineSimilarityNormalizedVecs(curCompVec, vector);
 				int indx = getBucketIndex(cosSimVal);
 				BitSet curBs = new BitSet(word2vec.size());
 				BitSet tempBs = csBucketContainer[i][indx];
-				if(tempBs!=null) {
+				if (tempBs != null) {
 					curBs.or(tempBs);
 				}
 				int temIndx = indx + 1;
@@ -199,10 +202,8 @@ public class W2VNrmlMemModelKMeans implements GenWord2VecModel {
 					finBitSet.and(curBs);
 				}
 			}
-			tl.printTime(1, "Setting Bits");
-			tl.logTime(1);
 			int nearbyWordsCount = finBitSet.cardinality();
-			LOG.info("Number of nearby words: "+nearbyWordsCount);
+			//LOG.info("Number of nearby words: " + nearbyWordsCount);
 			int[] nearbyIndexes = new int[nearbyWordsCount];
 			int j = 0;
 			for (int i = finBitSet.nextSetBit(0); i >= 0; i = finBitSet.nextSetBit(i + 1), j++) {
@@ -212,15 +213,12 @@ public class W2VNrmlMemModelKMeans implements GenWord2VecModel {
 					break; // or (i+1) would overflow
 				}
 			}
-			tl.printTime(1, "Extracting words");
-			tl.logTime(1);
 			closestWord = findClosestWord(nearbyIndexes, vector);
-			tl.printTime(1, "finding closest word");
 		} catch (Exception e) {
 			LOG.error("Exception has occured while finding closest word.");
 			e.printStackTrace();
 		}
-		LOG.info("Closest word found is: "+closestWord);
+		//LOG.info("Closest word found is: " + closestWord);
 		return closestWord;
 	}
